backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
sentence-transformers
faiss-cpu
pypdf
python-docx
python-multipart